+++ req/emails/bootstrap email gen text/account-changes-gentext.php
@@ -10,14 +10,14 @@
 	<div class="card mt-3" style="border-top:5px solid #007bff;">
       <img class="mx-auto mt-4 mb-3" width="42" height="30" src="https://hawg-ops.com/images/favicon.png"/>
       <div class="card-body">
-        <h3>Christopher,</h3>
+        <h3>__USERNAME__,</h3>
         <p class="mt-2">Changes to your account details were recently made. No action is required if you made this change.</p>
         <p class="mt-2">However, If you did not initiate this request, please contact us immediately by replying to this email.</p>
         <p class="mt-5">Thanks,</p>
         <p>Hawg Ops</p>
       </div>
   </div>
-  <div class="card mt-2">
+  <div class="card mt-2 mb-3">
     <div class="card-body">
       <a class="float-left"href="https://hawg-ops.com">Hawg Ops</a>
       <p class="float-right text-muted">Airmen Coders</p>