+++ req/emails/bootstrap email gen text/account-deleted-gentext.php
@@ -1,0 +1,34 @@
+<!DOCTYPE>
+<html>
+ <head>
+   <meta http-equiv="Content-Type" content="text/html; charset=utf-8" />
+   <link rel="stylesheet" media="all" href="/assets/application-mailer-dbc5154d3c4160e8fa7ef52fa740fa402760c39b5d22c8f6d64ad5999499d263.css" />
+ </head>
+ <!-- Edit the code below this line -->
+ <body class="bg-light">
+   <div class="container">
+	<div class="card mt-3" style="border-top:5px solid #007bff;">
+      <img class="mx-auto mt-4 mb-3" width="42" height="30" src="https://hawg-ops.com/images/favicon.png"/>
+      <div class="card-body">
+        <h3>__USERNAME__,</h3>
+        <p class="mt-2">It's been a long time since we've seen you, and we hate to see you go. But since your account has been inactive for over 15 months, it has been deleted.</p>
+        <p class="mt-5"><strong>Want to keep using Hawg Ops?</strong></p>
+        <p class="mt-2">No worries, you're always welcome back. You don't need an account to use the CAS planner - to save scenarios, you'll just have to copy/paste the data to a file on your computer.</p>
+        <p class="mt-5"><strong>Where are my scenarios?</strong></p>
+        <p class="mt-2">Below is your scenario information, if you have any that were saved to your account.</p>
+        <hr>
+        __SCENARIODATA__
+        <hr>
+        <p class="mt-2">If you have any questions, or need help with your account, feel free to contact Porkins at <a href="mailto:hawg.ops@gmail.com">hawg.ops@gmail.com</a>.</p>
+        <p class="mt-5">Thanks,</p>
+        <p>Hawg Ops</p>
+      </div>
+  </div>
+  <div class="card mt-2 mb-3">
+    <div class="card-body">
+      <a class="float-left"href="https://hawg-ops.com">Hawg Ops</a>
+      <p class="float-right text-muted">Airmen Coders</p>
+  </div>
+</div>
+</body>
+</html>