+++ req/emails/bootstrap email gen text/account-inactive-gentext.php
@@ -1,0 +1,35 @@
+<!DOCTYPE>
+<html>
+ <head>
+   <meta http-equiv="Content-Type" content="text/html; charset=utf-8" />
+   <link rel="stylesheet" media="all" href="/assets/application-mailer-dbc5154d3c4160e8fa7ef52fa740fa402760c39b5d22c8f6d64ad5999499d263.css" />
+ </head>
+ <!-- Edit the code below this line -->
+ <body class="bg-light">
+   <div class="container">
+	<div class="card mt-3" style="border-top:5px solid #007bff;">
+      <img class="mx-auto mt-4 mb-3" width="42" height="30" src="https://hawg-ops.com/images/favicon.png"/>
+      <div class="card-body">
+        <h3>__USERNAME__,</h3>
+        <p class="mt-2">We noticed you haven't used your Hawg Ops account __USEREMAIL__ in over a year.</p>
+        <p class="mt-5"><strong>Would you like to keep your account?</strong></p>
+        <p class="mt-2">If so, sign into Hawg Ops with __USEREMAIL__ within the next 3 months to keep your account active.</p>
+        <a class="btn btn-primary mt-5 mx-auto" href="https://hawg-ops.com/login">Sign in to keep your account</a>
+        <p class="mt-5"><strong>No longer want your account?</strong></p>
+        <p class="mt-2">No worries, we'll automatically disable your account 2 months from now, to give you one last chance, then delete your account 3 months from now.</p>
+        <p class="mt-5"><strong>What will happen to my scenarios?</strong></p>
+        <p class="mt-2">You can go into your account at any time to save your scenarios individually, then delete your account. Otherwise, before we delete your account, we will save and send you the scenario information to do with as you wish.</p>
+        <p class="mt-2">If you have any questions, or need help with your account, feel free to contact Porkins at <a href="mailto:hawg.ops@gmail.com">hawg.ops@gmail.com</a>.</p>
+        
+        <p class="mt-5">Thanks,</p>
+        <p>Hawg Ops</p>
+      </div>
+  </div>
+  <div class="card mt-2 mb-3">
+    <div class="card-body">
+      <a class="float-left"href="https://hawg-ops.com">Hawg Ops</a>
+      <p class="float-right text-muted">Airmen Coders</p>
+  </div>
+</div>
+</body>
+</html>