+++ req/emails/bootstrap email gen text/recover-account-gentext.php
@@ -1,0 +1,31 @@
+<!DOCTYPE>
+<html>
+ <head>
+   <meta http-equiv="Content-Type" content="text/html; charset=utf-8" />
+   <link rel="stylesheet" media="all" href="/assets/application-mailer-dbc5154d3c4160e8fa7ef52fa740fa402760c39b5d22c8f6d64ad5999499d263.css" />
+ </head>
+ <!-- Edit the code below this line -->
+ <body class="bg-light">
+   <div class="container">
+	<div class="card mt-3" style="border-top:5px solid #007bff;">
+      <img class="mx-auto mt-4 mb-3" width="42" height="30" src="https://hawg-ops.com/images/favicon.png"/>
+      <div class="card-body">
+        <h3>__USERNAME__,</h3>
+        <p class="mt-2">We received a request to recover your account on Hawg Ops. Click the button below to start the recovery process. Please act quick as your request is only valid for the next 30 minutes.</p>
+        <a class="btn btn-primary mt-5 mx-auto" href="https://hawg-ops.com/recover-account?email=__EMAIL__&token=__TOKEN__">Recover Account</a>
+        <p class="mt-5">Thanks,</p>
+        <p>Hawg Ops</p>
+        <p class="mt-5"><small>We recommend that you keep your password secure and not share it with anybody. If you feel that your password has been compromised, you can change it by going to your Account Settings.</small></p>
+        <p class="mt-2"><small>If you did not initiate this request, please contact us immediately by replying to this email.</small></p>
+        <hr>
+        <p class="mt-2"><small>If you're having trouble clicking the button above, copy and paste the following URL into your web browser: <a href="https://hawg-ops.com/recover-account?email=__EMAIL__&token=__TOKEN__">https://hawg-ops.com/recover-account?email=__EMAIL__&token=__TOKEN__</a></small></p>
+      </div>
+  </div>
+  <div class="card mt-2 mb-3">
+    <div class="card-body">
+      <a class="float-left"href="https://hawg-ops.com">Hawg Ops</a>
+      <p class="float-right text-muted">Airmen Coders</p>
+  </div>
+</div>
+</body>
+</html>