+++ req/emails/bootstrap email gen text/share-scenario-gentext.php
@@ -10,19 +10,19 @@
 	<div class="card mt-3" style="border-top:5px solid #007bff;">
       <img class="mx-auto mt-4 mb-3" width="42" height="30" src="https://hawg-ops.com/images/favicon.png"/>
       <div class="card-body">
-        <h3>Greetings,</h3>
-        <p class="mt-2">Christopher has shared a Hawg Ops CAS Scenario with you titled BAF East River Range. Please click on the button below to view the scenario.</p>
-        <a class="btn btn-primary mx-auto my-5" href="https://hawg-ops.com/cas?scenario=&share=1">View Scenario</a>
+        <h3>__SHAREDTO__,</h3>
+        <p class="mt-2">__SHAREDBY__ has shared a Hawg Ops CAS Scenario with you titled __SCENARIONAME__. Please click on the button below to view the scenario.</p>
+        <a class="btn btn-primary mx-auto my-5" href="https://hawg-ops.com/cas?scenario[]=__SCENARIOID__&share=1">View Scenario</a>
         <small>You must be logged into Hawg Ops to view this scenario. You will be directed to login or create an account as required.</small>
         <hr>
         <small>If you are having trouble trouble clicking the button, copy and paste the following URL into your web browser.</small>
-        <a href="https://hawg-ops.com/cas?scenario=&share=1"><small>https://hawg-ops.com/cas?scenario=&share=1</small></a>
+        <a href="https://hawg-ops.com/cas?scenario[]=__SCENARIOID__&share=1"><small>https://hawg-ops.com/cas?scenario[]=__SCENARIOID__&share=1</small></a>
                                                                                    
         <p class="mt-5">Thanks,</p>
         <p>Hawg Ops</p>
       </div>
   </div>
-  <div class="card mt-2">
+  <div class="card mt-2 mb-3">
     <div class="card-body">
       <a class="float-left"href="https://hawg-ops.com">Hawg Ops</a>
       <p class="float-right text-muted">Airmen Coders</p>